requirements.txt
absl-py==0.12.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
branca==0.4.2
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
cycler==0.10.0
dash==1.19.0
dash-auth==1.4.1
dash-bootstrap-components==0.12.0
dash-core-components==1.15.0
dash-daq==0.5.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Compress==1.9.0
Flask-SeaSurf==0.3.0
future==0.18.2
gunicorn==20.1.0
idna==2.10
ipykernel==5.5.0
ipyleaflet==0.13.6
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaleido==0.2.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matlabengineforpython===R2021a
matplotlib==3.4.1
mistune==0.8.4
Mosek==9.2.46
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
openpyxl==3.0.7
ortools==9.0.9048
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
peewee==3.14.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.17.2
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
retrying==1.3.3
rsome==0.1.3
scipy==1.6.3
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
terminado==0.9.3
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
traittypes==0.2.1
ua-parser==0.10.0
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
